+++ app/components/ui/game-button.test.tsx
@@ -5,7 +5,7 @@ import GameButton from "./game-button";
 
 vi.mock("next/image", () => ({
   __esModule: true,
-  default: (props: any) => {
+  default: ({ fill, priority, unoptimized, ...props }: any) => {
     const React = require("react");
     return React.createElement("img", props);
   },