+++ app/components/ui/game-icon.test.tsx
@@ -5,7 +5,7 @@ import GameIcon, { ICON_MAP, SIZE_PX } from "./game-icon";
 
 vi.mock("next/image", () => ({
   __esModule: true,
-  default: (props: any) => {
+  default: ({ fill, priority, unoptimized, ...props }: any) => {
     const React = require("react");
     return React.createElement("img", props);
   },