+++ app/forum/forum-icons.test.tsx
@@ -4,7 +4,7 @@ import { render } from "@testing-library/react";
 
 vi.mock("next/image", () => ({
   __esModule: true,
-  default: (props: any) => {
+  default: ({ fill, priority, unoptimized, ...props }: any) => {
     const React = require("react");
     return React.createElement("img", props);
   },